backend/requirements.txt
brotli==1.1.0
ConfigParser==7.0.0
dateparser==1.2.0
Django==4.2.13
djangorestframework==3.15.2
dl==0.1.0
docutils==0.17.1
HTMLParser==0.0.2
importlib_metadata==1.5.0
instaloader==4.12
jsonlines==4.0.0
keyring==18.0.1
lockfile==0.12.2
lxml==4.8.0
mysqlclient==2.2.4
numpy==1.21.4
ordereddict==1.1
protobuf==5.27.2
pyOpenSSL==24.1.0
python_dateutil==2.9.0.post0
pytz==2020.1
toml==0.10.2
tornado==6.4.1
typing_extensions==4.12.2
zipp==1.0.0
